feat: sign typed data v4
