+++ packages/sysweb3-keyring/package.json
@@ -34,9 +34,10 @@
     "crypto-js": "^4.1.1",
     "eth-chains": "^1.0.0",
     "eth-sig-util": "^3.0.1",
-    "ethereumjs-util": "^7.1.4",
+    "ethereumjs-util": "^7.1.5",
     "ethereumjs-wallet": "^1.0.2",
     "ethers": "^5.6.4",
+    "ethers-eip712": "^0.2.0",
     "graphql": "^16.3.0",
     "graphql-request": "^4.1.0",
     "lodash": "^4.17.21",
