--- packages/sysweb3-keyring/src/accounts/index.ts
@@ -1,1 +1,0 @@
-export * from './eth-accounts';
