+++ packages/sysweb3-keyring/src/index.ts
@@ -1,6 +1,7 @@
 export * from './keyring-manager';
-export * from './accounts';
+export * from './eth-manager';
 export * from './transactions';
 export * from './trezor';
 export * from './types';
+export * from './format';
 export * from './initial-state';
