+++ packages/sysweb3-keyring/src/transactions/index.ts
@@ -1,1 +1,2 @@
 export * from './syscoin';
+export * from './ethereum';
